fix: support for attribute expressioons in selections
