+++ query-service-impl/src/main/java/org/hypertrace/core/query/service/projection/ProjectionTransformation.java
@@ -2,23 +2,20 @@ package org.hypertrace.core.query.service.projection;
 
 import static io.reactivex.rxjava3.core.Single.zip;
 import static org.hypertrace.core.query.service.QueryRequestUtil.createBooleanLiteralExpression;
-import static org.hypertrace.core.query.service.QueryRequestUtil.createColumnExpression;
-import static org.hypertrace.core.query.service.QueryRequestUtil.createContainsKeyFilter;
 import static org.hypertrace.core.query.service.QueryRequestUtil.createDoubleLiteralExpression;
 import static org.hypertrace.core.query.service.QueryRequestUtil.createLongLiteralExpression;
 import static org.hypertrace.core.query.service.QueryRequestUtil.createNullNumberLiteralExpression;
 import static org.hypertrace.core.query.service.QueryRequestUtil.createNullStringLiteralExpression;
+import static org.hypertrace.core.query.service.QueryRequestUtil.createSimpleAttributeExpression;
 import static org.hypertrace.core.query.service.QueryRequestUtil.createStringLiteralExpression;
-import static org.hypertrace.core.query.service.QueryRequestUtil.isAttributeExpressionWithSubpath;
 
 import io.reactivex.rxjava3.core.Maybe;
 import io.reactivex.rxjava3.core.Observable;
 import io.reactivex.rxjava3.core.Single;
 import java.util.List;
 import java.util.Optional;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 import javax.inject.Inject;
+import lombok.extern.slf4j.Slf4j;
 import org.hypertrace.core.attribute.service.cachingclient.CachingAttributeClient;
 import org.hypertrace.core.attribute.service.projection.AttributeProjection;
 import org.hypertrace.core.attribute.service.projection.AttributeProjectionRegistry;
@@ -28,23 +25,16 @@ import org.hypertrace.core.attribute.service.v1.LiteralValue;
 import org.hypertrace.core.attribute.service.v1.Projection;
 import org.hypertrace.core.attribute.service.v1.ProjectionExpression;
 import org.hypertrace.core.attribute.service.v1.ProjectionOperator;
+import org.hypertrace.core.query.service.AbstractQueryTransformation;
 import org.hypertrace.core.query.service.QueryFunctionConstants;
 import org.hypertrace.core.query.service.QueryRequestUtil;
-import org.hypertrace.core.query.service.QueryTransformation;
 import org.hypertrace.core.query.service.api.AttributeExpression;
-import org.hypertrace.core.query.service.api.ColumnIdentifier;
 import org.hypertrace.core.query.service.api.Expression;
-import org.hypertrace.core.query.service.api.Filter;
 import org.hypertrace.core.query.service.api.Function;
-import org.hypertrace.core.query.service.api.Operator;
-import org.hypertrace.core.query.service.api.OrderByExpression;
-import org.hypertrace.core.query.service.api.QueryRequest;
 import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-final class ProjectionTransformation implements QueryTransformation {
-
-  private static final Logger LOG = LoggerFactory.getLogger(ProjectionTransformation.class);
+@Slf4j
+final class ProjectionTransformation extends AbstractQueryTransformation {
 
   private final CachingAttributeClient attributeClient;
   private final AttributeProjectionRegistry projectionRegistry;
@@ -57,96 +47,33 @@ final class ProjectionTransformation implements QueryTransformation {
   }
 
   @Override
-  public Single<QueryRequest> transform(
-      QueryRequest queryRequest, QueryTransformationContext transformationContext) {
-    return zip(
-            this.transformExpressionList(queryRequest.getSelectionList()),
-            this.transformExpressionList(queryRequest.getAggregationList()),
-            this.transformFilter(queryRequest.getFilter()),
-            this.transformExpressionList(queryRequest.getGroupByList()),
-            this.transformOrderByList(queryRequest.getOrderByList()),
-            (selections, aggregations, filter, groupBys, orderBys) ->
-                this.rebuildRequest(
-                    queryRequest, selections, aggregations, filter, groupBys, orderBys))
-        .doOnSuccess(transformed -> this.debugLogIfRequestTransformed(queryRequest, transformed));
+  protected Logger getLogger() {
+    return log;
   }
 
-  private Single<List<Expression>> transformExpressionList(List<Expression> expressionList) {
-    return Observable.fromIterable(expressionList)
-        .concatMapSingle(this::transformExpression)
-        .toList();
-  }
-
-  private Single<Expression> transformExpression(Expression expression) {
-    switch (expression.getValueCase()) {
-      case COLUMNIDENTIFIER:
-        return this.transformColumnIdentifier(expression.getColumnIdentifier());
-      case ATTRIBUTE_EXPRESSION:
-        return this.transformAttributeExpression(expression.getAttributeExpression());
-      case FUNCTION:
-        return this.transformFunction(expression.getFunction())
-            .map(expression.toBuilder()::setFunction)
-            .map(Expression.Builder::build);
-      case ORDERBY:
-        return this.transformOrderBy(expression.getOrderBy())
-            .map(expression.toBuilder()::setOrderBy)
-            .map(Expression.Builder::build);
-      case LITERAL:
-      case VALUE_NOT_SET:
-      default:
-        return Single.just(expression);
-    }
-  }
-
-  private Single<Expression> transformColumnIdentifier(ColumnIdentifier columnIdentifier) {
-    return this.projectAttributeIfPossible(columnIdentifier.getColumnName())
-        .map(expression -> this.aliasToMatchOriginal(getOriginalKey(columnIdentifier), expression))
-        .defaultIfEmpty(Expression.newBuilder().setColumnIdentifier(columnIdentifier).build());
-  }
-
-  private Single<Expression> transformAttributeExpression(AttributeExpression attributeExpression) {
+  @Override
+  protected Single<Expression> transformAttributeExpression(
+      AttributeExpression attributeExpression) {
     return this.projectAttributeIfPossible(attributeExpression.getAttributeId())
         .map(
-            expression ->
-                this.aliasToMatchOriginal(getOriginalKey(attributeExpression), expression))
+            projectedExpression ->
+                attributeExpression.hasSubpath()
+                    ? this.addSubpathOrThrow(projectedExpression, attributeExpression.getSubpath())
+                    : projectedExpression)
+        .map(expression -> this.aliasToMatchOriginal(attributeExpression, expression))
         .defaultIfEmpty(
             Expression.newBuilder().setAttributeExpression(attributeExpression).build());
   }
 
-  private Single<Function> transformFunction(Function function) {
-    return this.transformExpressionList(function.getArgumentsList())
-        .map(expressions -> function.toBuilder().clearArguments().addAllArguments(expressions))
-        .map(Function.Builder::build);
-  }
-
-  private Single<List<OrderByExpression>> transformOrderByList(
-      List<OrderByExpression> orderByList) {
-    return Observable.fromIterable(orderByList).concatMapSingle(this::transformOrderBy).toList();
-  }
-
-  private Single<OrderByExpression> transformOrderBy(OrderByExpression orderBy) {
-    return this.transformExpression(orderBy.getExpression())
-        .map(orderBy.toBuilder()::setExpression)
-        .map(OrderByExpression.Builder::build);
-  }
-
-  private Single<Filter> transformFilter(Filter filter) {
-    if (filter.equals(Filter.getDefaultInstance())) {
-      return Single.just(filter);
+  private Expression addSubpathOrThrow(Expression projectedExpression, String subpath) {
+    if (!QueryRequestUtil.isSimpleAttributeExpression(projectedExpression)) {
+      throw new IllegalArgumentException(
+          "Cannot use subpath for expression with non-trivial projection: " + projectedExpression);
     }
-
-    Single<Expression> lhsSingle = this.transformExpression(filter.getLhs());
-    Single<Expression> rhsSingle = this.transformExpression(filter.getRhs());
-    Single<List<Filter>> childFilterListSingle =
-        Observable.fromIterable(filter.getChildFilterList())
-            .concatMapSingle(this::transformFilter)
-            .toList();
-    return zip(
-        lhsSingle,
-        rhsSingle,
-        childFilterListSingle,
-        (lhs, rhs, childFilterList) ->
-            this.rebuildFilterOmittingDefaults(filter, lhs, rhs, childFilterList));
+    return Expression.newBuilder()
+        .setAttributeExpression(
+            projectedExpression.getAttributeExpression().toBuilder().setSubpath(subpath))
+        .build();
   }
 
   private Maybe<Expression> projectAttributeIfPossible(String attributeId) {
@@ -172,7 +99,8 @@ final class ProjectionTransformation implements QueryTransformation {
       Projection projection, AttributeKind expectedType) {
     switch (projection.getValueCase()) {
       case ATTRIBUTE_ID:
-        return this.transformExpression(createColumnExpression(projection.getAttributeId()));
+        return this.transformExpression(
+            createSimpleAttributeExpression(projection.getAttributeId()).build());
       case LITERAL:
         return this.rewriteLiteralAsQueryExpression(projection.getLiteral(), expectedType);
       case EXPRESSION:
@@ -280,13 +208,9 @@ final class ProjectionTransformation implements QueryTransformation {
     }
   }
 
-  private Expression aliasToMatchOriginal(String originalKey, Expression newExpression) {
+  private Expression aliasToMatchOriginal(AttributeExpression original, Expression newExpression) {
+    String originalKey = QueryRequestUtil.getAlias(original);
     switch (newExpression.getValueCase()) {
-      case COLUMNIDENTIFIER:
-        return newExpression.toBuilder()
-            .setColumnIdentifier(
-                newExpression.getColumnIdentifier().toBuilder().setAlias(originalKey))
-            .build();
       case ATTRIBUTE_EXPRESSION:
         return newExpression.toBuilder()
             .setAttributeExpression(
@@ -302,167 +226,5 @@ final class ProjectionTransformation implements QueryTransformation {
       default:
         return newExpression;
     }
-  }
-
-  private void debugLogIfRequestTransformed(QueryRequest original, QueryRequest transformed) {
-    if (!original.equals(transformed)) {
-      LOG.debug(
-          "Request transformation occurred. Original request: {} Transformed Request: {}",
-          original,
-          transformed);
-    }
-  }
-
-  /**
-   * This doesn't change any functional behavior, but omits fields that aren't needed, shrinking the
-   * object and keeping it equivalent to the source object for equality checks.
-   */
-  private Filter rebuildFilterOmittingDefaults(
-      Filter original, Expression lhs, Expression rhs, List<Filter> childFilters) {
-    Filter.Builder builder = original.toBuilder();
-
-    if (Expression.getDefaultInstance().equals(lhs)) {
-      builder.clearLhs();
-    } else {
-      builder.setLhs(lhs);
-    }
-
-    if (Expression.getDefaultInstance().equals(rhs)) {
-      builder.clearRhs();
-    } else {
-      builder.setRhs(rhs);
-    }
-
-    return builder.clearChildFilter().addAllChildFilter(childFilters).build();
-  }
-
-  private QueryRequest rebuildRequest(
-      QueryRequest original,
-      List<Expression> selections,
-      List<Expression> aggregations,
-      Filter originalFilter,
-      List<Expression> groupBys,
-      List<OrderByExpression> orderBys) {
-
-    QueryRequest.Builder builder = original.toBuilder();
-    Filter updatedFilter =
-        rebuildFilterForComplexAttributeExpression(originalFilter, orderBys, selections);
-
-    if (Filter.getDefaultInstance().equals(updatedFilter)) {
-      builder.clearFilter();
-    } else {
-      builder.setFilter(updatedFilter);
-    }
-
-    return builder
-        .clearSelection()
-        .addAllSelection(selections)
-        .clearAggregation()
-        .addAllAggregation(aggregations)
-        .clearGroupBy()
-        .addAllGroupBy(groupBys)
-        .clearOrderBy()
-        .addAllOrderBy(orderBys)
-        .build();
-  }
-
-  /*
-   * We need the CONTAINS_KEY filter in all filters, selections and order bys dealing with complex
-   * attribute expressions as Pinot gives error if particular key is absent. Rest all work fine.
-   * To handle order bys and selections, we add the corresponding filter at the top and 'AND' it with the main filter.
-   * To handle filter, we modify each filter (say filter1) as : "CONTAINS_KEY AND filter1".
-   */
-  private Filter rebuildFilterForComplexAttributeExpression(
-      Filter originalFilter, List<OrderByExpression> orderBys, List<Expression> selections) {
-
-    Filter updatedFilter = updateFilterForComplexAttributeExpressionFromFilter(originalFilter);
-    List<Filter> filterList =
-        Stream.concat(
-                createFilterForComplexAttributeExpressionFromOrderBy(orderBys),
-                createFilterForComplexAttributeExpressionFromSelection(selections))
-            .collect(Collectors.toList());
-
-    if (filterList.isEmpty()) {
-      return updatedFilter;
-    }
-
-    if (!updatedFilter.equals(Filter.getDefaultInstance())) {
-      return Filter.newBuilder()
-          .setOperator(Operator.AND)
-          .addChildFilter(updatedFilter)
-          .addAllChildFilter(filterList)
-          .build();
-    }
-
-    if (filterList.size() > 1) {
-      return Filter.newBuilder().setOperator(Operator.AND).addAllChildFilter(filterList).build();
-    } else {
-      return filterList.get(0);
-    }
-  }
-
-  private Filter updateFilterForComplexAttributeExpressionFromFilter(Filter originalFilter) {
-    /*
-     * If childFilter is present, then the expected operators comprise the logical operators.
-     * If childFilter is absent, then the filter is a leaf filter which will have lhs and rhs.
-     */
-    if (originalFilter.getChildFilterCount() > 0) {
-      Filter.Builder builder = Filter.newBuilder();
-      builder.setOperator(originalFilter.getOperator());
-      originalFilter
-          .getChildFilterList()
-          .forEach(
-              childFilter ->
-                  builder.addChildFilter(
-                      updateFilterForComplexAttributeExpressionFromFilter(childFilter)));
-      return builder.build();
-    } else if (isAttributeExpressionWithSubpath(originalFilter.getLhs())) {
-      Filter childFilter =
-          createContainsKeyFilter(originalFilter.getLhs().getAttributeExpression());
-      return Filter.newBuilder()
-          .setOperator(Operator.AND)
-          .addChildFilter(originalFilter)
-          .addChildFilter(childFilter)
-          .build();
-    } else {
-      return originalFilter;
-    }
-  }
-
-  private Stream<Filter> createFilterForComplexAttributeExpressionFromOrderBy(
-      List<OrderByExpression> orderByExpressionList) {
-    return orderByExpressionList.stream()
-        .map(OrderByExpression::getExpression)
-        .filter(QueryRequestUtil::isAttributeExpressionWithSubpath)
-        .map(Expression::getAttributeExpression)
-        .map(QueryRequestUtil::createContainsKeyFilter);
-  }
-
-  private Stream<Filter> createFilterForComplexAttributeExpressionFromSelection(
-      List<Expression> selections) {
-    return selections.stream()
-        .flatMap(this::getAnyAttributeExpression)
-        .map(QueryRequestUtil::createContainsKeyFilter);
-  }
-
-  private Stream<AttributeExpression> getAnyAttributeExpression(Expression selection) {
-    if (selection.hasFunction()) {
-      return selection.getFunction().getArgumentsList().stream()
-          .flatMap(this::getAnyAttributeExpression);
-    } else {
-      return Stream.of(selection)
-          .filter(QueryRequestUtil::isAttributeExpressionWithSubpath)
-          .map(Expression::getAttributeExpression);
-    }
-  }
-
-  private String getOriginalKey(AttributeExpression attributeExpression) {
-    String alias = attributeExpression.getAlias();
-    return alias.isEmpty() ? attributeExpression.getAttributeId() : alias;
-  }
-
-  private String getOriginalKey(ColumnIdentifier columnIdentifier) {
-    String alias = columnIdentifier.getAlias();
-    return alias.isEmpty() ? columnIdentifier.getColumnName() : alias;
   }
 }
